% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlleleFreq.R
\name{calculateAlleleFreq}
\alias{calculateAlleleFreq}
\title{Calculate the allele frequency of each variant in a VCF file}
\usage{
calculateAlleleFreq(vcf)
}
\arguments{
\item{vcf}{A data frame containing the VCF file read by the read.vcf function \code{\link{read.vcf}}}
}
\description{
Calculate the allele frequency of each variant in a VCF file
}
\examples{
\dontrun{
library(vcfPIC)
vcfData<-readVCF("data/sheep_genotypes.vcf")
calculateAlleleFreq(vcfData)
}
}
