% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlleleFreqBinary.R
\name{calculateAlleleFreqBinary}
\alias{calculateAlleleFreqBinary}
\title{Calculate allele frequency for Binary file read by \code{\link{readGeneticBinaryTable}} function}
\usage{
calculateAlleleFreqBinary(BinaryTable, roundnum = 2)
}
\arguments{
\item{BinaryTable}{A data frame containing Binary data}
}
\value{
A data frame containing allele frequency
}
\description{
Calculate allele frequency for BinaryTable  file read by \code{\link{readGeneticBinaryTable}} function
}
\examples{
binaryData<-readGeneticBinaryTable("data/sheep_genotypes_binary.tsv")
calculateAlleleFreqBinary(binaryData)
}
