% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlleleFreqHapMap.R
\name{calculateAlleleFreqHapMap}
\alias{calculateAlleleFreqHapMap}
\title{Calculate allele frequency for haplotype  file}
\usage{
calculateAlleleFreqHapMap(hapmapData, roundnum = 2)
}
\arguments{
\item{hapmapData}{A data frame containing haplotype data}
}
\value{
A data frame containing allele frequency
}
\description{
Calculate allele frequency for haplotype  file read by \code{\link{readHapmap}} function
}
\examples{
hapmapData<-readHapmap("data/sample_hapDouble.hapmap")
calculateAlleleFreqHapMap(hapmapData)
}
