% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlleleFreqVCF.R
\name{calculateAlleleFreqVCF}
\alias{calculateAlleleFreqVCF}
\title{Calculate the allele frequency of each variant in a VCF file}
\usage{
calculateAlleleFreqVCF(vcf, roundnum = 2)
}
\arguments{
\item{vcf}{A data frame containing the VCF file read by the read.vcf function \code{\link{read.vcf}}}

\item{roundnum}{The number of decimal places to round the output to}
}
\description{
Calculate the allele frequency of each variant in a VCF file
}
\examples{
\dontrun{
library(vcfPIC)
vcfData<-readVCF("data/sheep_genotypes.vcf")
calculateAlleleFreq(vcfData)
}
}
