% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIC.R
\name{calculatePIC}
\alias{calculatePIC}
\title{Calculate Polymorphic Information Content (PIC) from VCF , Hapmap file, Genetic Binary Table read by \code{\link{readGeneticBinaryTable}}}
\usage{
calculatePIC(filepath, filetype, outpath = "")
}
\arguments{
\item{filepath}{path to the file}

\item{filetype}{type of file, either vcf , hapmap , or genetic binary table}

\item{outpath}{path to write the PIC values}
}
\value{
PIC values
}
\description{
Calculate Polymorphic Information Content (PIC) from VCF , Hapmap file, Genetic Binary Table read by \code{\link{readGeneticBinaryTable}}
}
\examples{
calculatePIC("data/sheep_genotypes.vcf", "vcf")
calculatePIC("data/sheep_genotypes.hmp", "hapmap")
calculatePIC("data/sheep_genotypes_binary.tsv", "binary")
}
