% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePICByFreq.R
\name{calculatePICByFreq}
\alias{calculatePICByFreq}
\title{Calculate the polymorphism information content (PIC) of a set of alleles from allele frequencies}
\usage{
calculatePICByFreq(AlleleFreq)
}
\arguments{
\item{AlleleFreq}{A matrix of allele frequencies,
with rows representing alleles and columns representing individuals \code{\link{calculateAlleleFreq}}}
}
\value{
The polymorphism information content (PIC) of the alleles
}
\description{
Calculate the polymorphism information content (PIC) of a set of alleles from allele frequencies
}
