% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGenoBi2Numeric.R
\name{convertGenoBi2Numeric}
\alias{convertGenoBi2Numeric}
\title{Convert a habmap object to a numeric matrix}
\usage{
convertGenoBi2Numeric(habmap)
}
\arguments{
\item{habmap}{a habmap object}
}
\value{
a numeric matrix
}
\description{
Convert a habmap object to a numeric matrix
}
