% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIC.R
\name{PICvec}
\alias{PICvec}
\title{Calculate the polymorphism information content (PIC) of a set of alleles}
\usage{
PICvec(freq)
}
\arguments{
\item{freq}{A numeric vector of allele frequencies}
}
\value{
The polymorphism information content (PIC) of the alleles
}
\description{
Calculate the polymorphism information content (PIC) of a set of alleles
}
