% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeneticBinaryTable.R
\name{readGeneticBinaryTable}
\alias{readGeneticBinaryTable}
\title{Read a binary table of genetic data coded as 0, 1, 2, -1}
\usage{
readGeneticBinaryTable(binartablepath, header = TRUE, sep = "\\t", ...)
}
\arguments{
\item{binartablepath}{The path to the binary table}

\item{header}{A logical value indicating whether the file has a header line}

\item{sep}{The field separator character}
}
\value{
A data frame with the binary table
}
\description{
Read a binary table of genetic data coded as 0, 1, 2, -1
Where AA = 0, AB = 1, BB = 2, and missing data = -1
}
\examples{
readBinaryTable("data/sheep_genotypes_binary.tsv")
}
