% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHapmap.R
\name{readHapmap}
\alias{readHapmap}
\title{Read hapmap file into R}
\usage{
readHapmap(path, sep = "\\t", header = TRUE, ...)
}
\arguments{
\item{path}{Path to hapmap file}

\item{sep}{Separator}

\item{header}{Logical indicating if the file has a header}

\item{...}{Additional arguments to \code{\link{read.table}}}
}
\value{
A data frame
}
\description{
Read hapmap file into R
}
\examples{
readHapmap("hapmap.txt")
readHapmap("hapmap.txt", header = FALSE)
readHapmap("hapmap.txt", header = FALSE, sep = "\t")
readHapmap("hapmap.txt", header = FALSE, sep = "\t", skip = 1)
readHapmap("hapmap.txt", header = FALSE, sep = "\t", skip = 1, nrows = 10)
}
