% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep_genotypes}
\alias{sheep_genotypes}
\title{Data of Sheep Genotyping
VCF file containing the genotypes of 160 Sheeps}
\format{
Data in VCF format
}
\source{
Aboul-Naga, A.M., Alsamman, A.M., El Allali, A., Elshafie, M.H., Abdelal, E.S., Abdelkhalek, T.M., Abdelsabour, T.H., Mohamed, L.G. and Hamwieh, A., 2022. Genome-wide analysis identified candidate variants and genes associated with heat stress adaptation in Egyptian sheep breeds. Frontiers in Genetics, 13, p.898522.
}
\usage{
sheep_genotypes
}
\description{
Data of Sheep Genotyping
VCF file containing the genotypes of 160 Sheeps
}
\examples{
data(sheep_genotypes)
}
\keyword{datasets}
